Cleanup warnings
